Server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
ascii_colors==0.11.6
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
configparser==7.2.0
cors==1.0.1
distro==1.9.0
fastapi==0.124.0
filelock==3.20.1
frozenlist==1.8.0
future==1.0.0
gevent==25.9.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
json_repair==0.54.3
lxml==6.0.2
multidict==6.7.0
nano-vectordb==0.0.4.3
neo4j==6.0.3
networkx==3.6.1
numpy==1.26.4
openai==2.14.0
packaging==25.0
pandas==2.3.3
pipmaster==1.0.10
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
pypinyin==0.55.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tldextract==5.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.22.0
zope.event==6.1
zope.interface==8.1.1
